+++ packages/eos-transit-ledger-provider/patch-Transportwebauthn.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -x
+cd "$(dirname "$0")";
+echo "Patching TransportWebAuthn.js ..."
+patch node_modules/@ledgerhq/hw-transport-webauthn/lib/TransportWebAuthn.js Transportwebauthn.js.patch